refactor: change place of Navigation
